Projeto utilizado no curso de PHP Composer: Dependências, Autoload e Publicação
