--- composer-aula6-final/README.md
@@ -1,3 +1,0 @@
-# Documentação do componente
-
-Este componente é SUPIMPA!!
