--- composer-aula6-final/composer.json
@@ -1,47 +1,0 @@
-{
-    "name": "cviniciussdias/buscador-cursos",
-    "description": "Projeto que busca os cursos no site da Alura",
-    "type": "library",
-    "authors": [
-        {
-            "name": "Vinicius Dias",
-            "email": "carlosv775@gmail.com"
-        }
-    ],
-    "require": {
-        "guzzlehttp/guzzle": "^6.3",
-        "symfony/dom-crawler": "^4.2",
-        "symfony/css-selector": "^4.2"
-    },
-    "autoload": {
-        "files": [
-            "functions.php"
-        ],
-        "psr-4": {
-            "Alura\\BuscadorDeCursos\\": "src/"
-        }
-    },
-    "require-dev": {
-        "phpunit/phpunit": "^8.1",
-        "squizlabs/php_codesniffer": "^3.4",
-        "phan/phan": "^1.3"
-    },
-    "scripts": {
-        "test": "phpunit tests\\TestBuscadorDeCursos.php",
-        "cs": "phpcs --standard=PSR12 src/",
-        "phan": "phan --allow-polyfill-parser",
-        "check": [
-            "@phan",
-            "@cs",
-            "@test"
-        ],
-        "post-update-cmd": [
-            "@test"
-        ]
-    },
-    "scripts-descriptions": {
-        "check": "Roda as verificações do código. PHAN, PHPCS e PHPUNIT"
-    },
-    "license": "GPL-3.0",
-    "bin": ["buscar-cursos.php"]
-}
