--- composer-aula6-final/functions.php
@@ -1,6 +1,0 @@
-<?php
-
-function exibeMensagem(string $mensagem)
-{
-    echo $mensagem . PHP_EOL;
-}
