+++ php/php-composer-dependencias-autoload-e-publicacao/.phan/config.php
@@ -1,0 +1,22 @@
+<?php
+
+return [
+    "target_php_version" => '7.3',
+    'directory_list' => [
+        'src',
+        'vendor/symfony/dom-crawler',
+        'vendor/guzzlehttp/guzzle',
+        'vendor/psr/http-message'
+    ],
+    "exclude_analysis_directory_list" => [
+        'vendor/'
+    ],
+    'plugins' => [
+        'AlwaysReturnPlugin',
+        'UnreachableCodePlugin',
+        'DollarDollarPlugin',
+        'DuplicateArrayKeyPlugin',
+        'PregRegexCheckerPlugin',
+        'PrintfCheckerPlugin',
+    ],
+];
