+++ php/php-composer-dependencias-autoload-e-publicacao/README.md
@@ -1,0 +1,14 @@
+# PHP Composer: Dependências, Autoload e Publicação
+
+Categoria | Carga horária | Atividades | Finalizado | Certificado |
+:-:|:-:|:-:|:-:|:-:|
+Programação | 6 horas | 58 | 24/03/2022 | [Certificado Normal](https://cursos.alura.com.br/certificate/94c93fff-6e45-43dd-8c8d-3b6921a7089c) / [Certificado Formal](https://cursos.alura.com.br/user/rodineicosta/course/php-composer/formalCertificate)
+
+## Aulas
+
+- Instalando o Composer
+- Gerenciando dependências
+- Entendendo Autoload
+- Ferramentas de qualidade de código
+- Automatizando processos com Scripts
+- Publicando um pacote
