+++ php/php-composer-dependencias-autoload-e-publicacao/buscar-cursos.php
@@ -1,0 +1,18 @@
+#!/usr/bin/env php
+<?php
+
+require 'vendor/autoload.php';
+
+use Alura\BuscadorDeCursos\Buscador;
+use GuzzleHttp\Client;
+use Symfony\Component\DomCrawler\Crawler;
+
+$client = new Client(['base_uri' => 'https://www.alura.com.br/']);
+$crawler = new Crawler();
+
+$buscardor = new Buscador($client, $crawler);
+$cursos = $buscardor->buscar('/cursos-online-programacao/php');
+
+foreach ($cursos as $curso) {
+    exibeMensagem($curso);
+}
