+++ php/php-composer-dependencias-autoload-e-publicacao/functions.php
@@ -1,0 +1,6 @@
+<?php
+
+function exibeMensagem(string $mensagem)
+{
+    echo $mensagem . PHP_EOL;
+}
