+++ php/php-composer-dependencias-autoload-e-publicacao/src/Buscador.php
@@ -1,0 +1,41 @@
+<?php
+
+namespace Alura\BuscadorDeCursos;
+
+use GuzzleHttp\ClientInterface;
+use Symfony\Component\DomCrawler\Crawler;
+
+class Buscador
+{
+    /**
+     * @var ClientInterface
+     */
+    private $httpClient;
+    /**
+     * @var Crawler
+     */
+    private $crawler;
+
+    public function __construct(ClientInterface $httpClient, Crawler $crawler)
+    {
+        $this->httpClient = $httpClient;
+        $this->crawler = $crawler;
+    }
+
+    public function buscar(string $url): array
+    {
+        $resposta = $this->httpClient->request('GET', $url);
+
+        $html = $resposta->getBody();
+        $this->crawler->addHtmlContent($html);
+
+        $elementosCursos = $this->crawler->filter('span.card-curso__nome');
+        $cursos = [];
+
+        foreach ($elementosCursos as $elemento) {
+            $cursos[] = $elemento->textContent;
+        }
+
+        return $cursos;
+    }
+}
